turn markdown linebreak into `<w:br />` and add <w:p> style.
Previously, linebreaks were translated into a new paragraph.
This made it impossible to apply different styling rules to
new paragraphs but not linebreaks.

Every paragraph now uses the style `Paragraph` if the style
is not available Word falls back to `Normal`. This makes it possible
to configure a style only for Sablon paragraphs.