implement markdown insertion using html insertion.
The previously packaged wordml formatter for redcarpet
`Sablon::Redcarpet::Render::WordML` has severe shortcomings. It has been
removed in favor of the newly added HTML insertion. The
`Sablon::Content::Markdown` is now implemented on top of
`Sablon::Content::HTML`. It will stay around for a bit but eventuelly
the responsability to turn the markdown into HTML will be placed on
client applications.