Update html_sample fixture with corrected behavior.
Due to the changes I defer converting the CSS to OpenXML until I actually
instantiate an AST node. That means when working with "pass-through" HTML
elements that only apply a style the CSS properties get thrown around instead
of WordML properties. Preiously I converted the CSS to WordML immediately and
and in the case of text-decoration the WordML produced from the conversion was not
always the same so it didn't get overridden properly i.e. line-through produces a
w:strike while underline produces a w:u. Now due to the fact I defer conversion
text-decoration will be overridden correctly as it remains 'text-decoration' until
the Run AST class itself is created.