Implement config module (#67)
* Create initial version of config module to register HTML tags

* Add tests for Configuration::HTMLTag

* Fully test the configuration module

* Minor changes to configuration files and ensure allowed_children is always array of symbols

* Create initial methods in converter utilizing the config and added tests for them

The methods are currently unused to allow easier testing

* convert properties to strings as final step in NodeProperties#filter

* Add document fragment as root level tag in configuration

This greatly simplifies logic when assessing parent-child validity

* Fix unit tests in converter_test

* Additional fixes to html element initialization.

my ul and ol definitions accidentally had pstyle instead of pStyle.
I also didn't need to call .to_sym on the doc fragment key. It gets
converted to a symbol during HTMLTag initialization

* Refactor HTML converter to use the new config module

* Prevent p and div from being nested inside an inline tag

* Update config to not allow mixed lists

* Require all children of ol or ul tags to be an li tag or a deeper nesting of the parent tag

* WIP: Create initial AST classes for a List(ul & ol) and a ListParagraph(li)

This classes also define the new interface I am going to design for all AST
nodes. This commit by itself does not affect normal code execution

* Standardize all ast node initialization methods. #ci skip

This standardization provides a lot of flexiblity while keeping
everything consistent. It would also be a necessity if I allow the user
to define their own AST classes.

* Begin work on a revamped HTML to AST converter

* Update default config to use new List and ListParagraph AST nodes

* Update ast builder to properly handler node properties

I previously wasn't getting styles applied correctly from pass-through nodes

* Remove all extraneous code from the HTML convert class.

Most of this logic has ben refactored into the ast_builder and ast node
sub classes.

* Refactor AST node classes to handle styles more effectively in the new system.

The largest change is that lists are handled within the AST class instead of
inside the converter itself

* Don't convert keys to symbols inside NodeProperties#filter

* Move node properties testing into a separate file

The single converter_test file was becoming extremely cluttered and
hard to work with. I also fixed test_transferred_properties to use the
new instance method instead of a class method

* Refactor NodeProperties class to setup transferred properties on initialization

* Update node properties test to check transferred_properties instance var

* Properly handled nested lists so they share the same numId

Also added an inspect method to the List ast class and fixed the
inspect method in the Paragraph ast class

* Move the AST test class into it's own file. Also fixed the get_numpr_prop method

* Fix List AST node so ListParagraphs not longer get nested within each other

Styles also appear to properly transfer

* Fix recursive method convert_style_property to exit when the Node class is encountered

* Create a test file for the ast_builder (incomplete)

[ci skip]

* Update html_sample fixture with corrected behavior.

Due to the changes I defer converting the CSS to OpenXML until I actually
instantiate an AST node. That means when working with "pass-through" HTML
elements that only apply a style the CSS properties get thrown around instead
of WordML properties. Preiously I converted the CSS to WordML immediately and
and in the case of text-decoration the WordML produced from the conversion was not
always the same so it didn't get overridden properly i.e. line-through produces a
w:strike while underline produces a w:u. Now due to the fact I defer conversion
text-decoration will be overridden correctly as it remains 'text-decoration' until
the Run AST class itself is created.

* Madify style hash order so inline styles can genreally override tag styles.

This is only the case for non-toggle properties or if the toggle property has a value that can be
set to false. i.e. <w:u w:val="false" />. Otherwise the style will persist because
CSS can only override not remove.

* Strip whitespace from style properties after splitting the style string

* Finish testing the ASTBuilder

The remainder of the code is excercised by converter.rb

* Move all AST class CSS style conversions from class constants to a config instance var

This instance variable is accessed from the class level method style_conversion,
which can be overriden in subclasses as desired. This allows the end user
to add their own AST nodes without having to separately register their
styles with the config instance.

* Add methods to the configration object to add and remove CSS style conversions

* Add a style converter for sz

The same string to numeric conversion is used in border and font-size
this pseduo key ensures that things are done consistently.

* Add test to ensure newly registered tags can be converted

* Add a test that converts a registered HTML node that provides it's own AST class

* Add test to ensure registered style converters are successfully used

* Update configuration_test to git the register and remove style converter methods

* Add initial HTML tag conversion docs

* Tweak comments and add content to section describing CSS style customizations

* Fix grammar and readability in new Config section of README

* Add Table of contents to README

Generated using github-markdown-toc.
https://github.com/ekalinin/github-markdown-toc

* Fix typo in README [ci skip]