Give subclasses more flexibility to define child content of nodes.
Previously any children needed to implement the `to_docx` method which
lead to alot of structs wrapping plain text content. Now with the
`children_to_docx` method a subclass can delegate that directly to it's
child nodes as with Paragraph, `runs.to_docx` or specify plain text directly
as the Run and Newline classes do.