Fix bug when the user registers a new HTML tag
If the user registers a new HTML tag and specifies style properties
for the element using string keys the convert will attempt to process
them as if they came from the `style=` attribute. For non-string
values this causes an error. The fix was to onvert all property keys
specified during registration to symbols.