Use `to_s` instead of `text` in the Run AST node (#80)
Allows HTML entities encountered in text nodes to be properly re-encoded during the HTML to WordML conversion process. Otherwise when the WordML syntax gets re-parsed as XML certain characters get scrubbed (i.e. the less than and ampersand characters)

* Use `to_s` instead of `text` in the Run AST node

When using `text` the string is returned as we would want to
*see* it, i.e. "&amp;" is output as "&". Using `to_s` re-encodes
the entities or in a sense leaves them "as-is". This is needed for
HTML conversion since it gets parsed as HTML, then converted to
static XML (MS Word markup) which finally gets reparsed by Nokogiri
prior to final output into the document nodeset. This means entities
can get lost. See issue #79.

* Add test case to show entities get displayed correctly

* Update HTML insertion test fixture