Move NodeProperties out of html/ast.rb and into it's own file
the ast.rb file was getting very long, over 500 lines and all of
the other classes within are actual AST nodes. Moving it out of
that file makes it clear the NodeProperties class is not an actual
AST node and instead just a supporting object used by them. It also
makes the file slightly more managable by reducing the length.