Add the XML tags permitted by the OOXML spec for table cells and paragraphs
Table cells can accept a wide variety of elements, more than we want
to try and anticipate. So instead exhaustive whitelists have been added.
The difference of these tags can be used to identify elements that
should be wrapped by a paragraph tags. Tags that aren't in either
list are ignored and MS Word is left to handle the discrepancy.