Add an #== method to the WordML content type
This change was necessary because I changed the xml attribute from being
a plain text string to a live Nokogiri document fragment. The overridden
instances. When they are compared the xml attribute's of each are compared
as strings for equality instead of checking the entire object itself. This
was mainly needed for testing but can be used as needed in production code.