Create initial version of new docx equality assertion
I now need to inspect more files than just the document.xml since
other content such as headers, footers, and *.rels files may be
modified during the run. Additionally, moving the assertion into
the Minitest::Assertions module gives me access to a tool to show
a diff of the extracted XML. This is useful for small changes in files.