Implement dom for templates (#85)
Implements a very basic document object model that allows for interactions between multiple files to be defined cleanly. Additional file handlers can be added as needed by the end user to extend functionality.

* Temporarily commit additional development dependencies

* Create very simple DOM model class

This class currently just stores the entire template in memory and
provides no additional abstraction or helper methods to ease interactions.

* Generalize call signature of existing processors

This is needed for changes being made to how the template is
processed.

* Update relationship to return the document root after modifying inplace

Previously it would actually write to the zip file output stream
directly. This was needed to work with the old way templates were
being processed.

* Refactor Sablon::Template to work with DOM and process files non-sequentially.

* Rename dom.rb to model.rb

* Add a #document method to the Environment instance that calls Template#document

* Track the current entry during processing of the template

* Converted the relationships class into a DOM helper class

* Add an .extend_model method to the relationships handler

This allows it to add methods to the DOM Model class to clean up
logic at call sites and avoid polluting the Model class with a ton
of helpers.

* Remove the old requrie relationships line

* Add logic to teh DOM model class to register content handlers

These handlers can then extend the Model class.

* Remove the relationship accessor from the environemnt instance

* Update teh Hyperlink AST class to use the new DOM based method to add a relationship

* Remove special logic from Template when handling the document.xml.rels file

* Create a base class for all file handlers to inheirit from

* Use FileHandler super class with Relationships handler

* Prevent modification during iteration error in Template#process

This can occur if a new entry is added to the zip contents hash
while it is being processed. The prime example being a new *.rels
file getting created.

* Initialize @document during #render so a template can be reused.

Otherwise since the zip contents gets modified in place only the first
call to #render would have an effect.

* Generalize file handler init arg to just be "content"

In theory any type of content could be handled by a file hander
not just XML files.

* Add #create_entry_if_not_exist method to Sablon::DOM::Model

This offers an easy route to create new entries in the output zip file
when needed. (i.e. creation of a new *.rels entry for say the header).

* Allow new rels files to be created if they don't already exist.

* Add classes to test_helper to serve as mock Templates and DomModels

* Fix unit tests to work with new MockTemplate and MockDomModel

This replaces the use of the UUIDTestGenerator in those test cases

* Remove 'require secrurerandom' from ast_test

I don't use it anymore.

* Remove UID test logic from html insertion integration test

* Ensure I don't try to read a directory in the zip file during DOM::Model#initalize

* Remove old docx equal assertion

* Create initial version of new docx equality assertion

I now need to inspect more files than just the document.xml since
other content such as headers, footers, and *.rels files may be
modified during the run. Additionally, moving the assertion into
the Minitest::Assertions module gives me access to a tool to show
a diff of the extracted XML. This is useful for small changes in files.

* Refactor docx equal assertion to normalize XML and output a diff

I also changed it so it only checks XML files, everything not ending
in .xml or .rels is ignored.

* Update recipe sample fixture so all files match

The doument.xml file matched fine but some of the other files such as
numbering had leftover XML from previous edits.

* Update HTML sample test fixture due to change in how rId's are added

* Remove the old numbering class and numbering.xml processor

This file doesn't need to be processed separately anymore like it
was before because using the new DOM logic numbering definitions
are added on the fly.

* Create a new DOM file handler specifically for numbering.xml

This takes the place of the numbering class that got registered
to the environment as well as the numbering processor used to add
new definitions to numbering.xml

* Update the List AST class to use new method to register a list

* Update docx integration test fixtures to match new list numid's

The numid's are now sequential instead of starting off at an arbitrary
value of 1000 so all of the test fixtures needed updated to compensate
for the change.

* Add numbering logic to the MockDom test helper class

* Fix all unit tests to use the new DOM method to work with numbering.

* Revert "Temporarily commit additional development dependencies"

This reverts commit 5b619a90dfa1269274639af705e4778cfddaf105.

* Remove logic that allowed passing in of a parent environment to Environment#initialize

This logic was needed because I had to pass in things like numbering and
relationships directly but those are now handled by the DOM which is
pulled from the template.

* Remove numbering processor elif clause in Template#get_processor

* Fix incorrect information in the README

* Add a #max_attribute_value method to the FileHandler class

This provides an easy way to get the maximum value of an attribute
for a given nodeset and selector.

* Wrap all entries with the FileHandler class unless one is specified

This prevents an odd mix of raw Nokogiri XML and FileHandler subclasses
from floating around together in the DOM. It also means #max_attribute_value
can be called on any file in the DOM (assuming it is XML).