Implement document processor registration interface (#86)
Implements an interface where additional processor classes can be added externally without requiring a fork of the gem or awkward monkey patching. Example usage of processor registration can be seen in lib/sablon/template.rb

* Add logic to register new processors

* Refactor Template to use new processor registration logic

The #get_processor method was moved to the Template class
and is called for every file in the zip file, that loop iteration
is skipped if a processor doesn't exist yet.

* Remove *args logic from processor classes

This was needed before because I sent a variable number of
arguments based on the processor type. I don't need to do that
anymore.

* Store section propeties hash as an attrbiute on the Environment instance

This will allow me to remove it from the args passed into specific
processors since it is only needed by the Document processor and
more specifically only needed by it when processing the document.xml
file.

* Allow multiple processors to be registered to the same pattern

* Plurialize #get_processor method to reflect it now returns more than one.

* Refactor Processor::SectionProperties into a stand alone processor class

Previously it used the Processor::Document class to set it up and then
edit the document. The actual functionality of the processor is unchanged.

* Readd start_page_number getter and setter methods to Processor::SectionProperties

I'd be curious if they were ever used directly but it doesn't add any
overhead to keep them around since the tests are already written.

* Add #section_properties= setter method to Environment class

This was added to ensure the properties hash being provided to
the evironment is always modified. This is only done so all keys
are strings to simplfy call site access.

* Modify SectionPropertiesTest to work with changes to that processor class

* Move section properties test file to test/processors

This is where it actually belongs beacuse it is testing a
processor method.

* Remove unused get_processors method from Template