Fix unique ids in loops (#89)
Uses the new DOM logic to prevent document corruption when shapes and images are present in a loop construct. This PR also fixes a backwards compatibility issue that arose in HTML insertion when plain text was a top level node in the HTML being parsed.

* Allow DOM entries to be accessed by name using []

* Add logic to Loop operation to update unique docPr and cNvPr tags

* Update Loop operation to not require namespaces for docPr and cNvPr tags

These can exist under multiple namespaces and while the id attribute
probably only has to be unique per namespace it is far simpler to update
them all in unison. If this is found to break something then we will address
the issue then.

* Update mock DOM class to work with new unique ID logic in Loop operation

* Add a unit test for unique id incrementing inside loops

* Fix duplicate class name in sablon_test

This prevented the CV test case from executing, a test case that now fails due
to changes to HTML insertion.

* Fix test failure due to header text being changed to current time

This wasn't checked prior to my changes to checking docx equality
since the old method only checked the document.xml instead of every
XML file in the archive.

* Fix CV test case for Sablon, HTML content needed wrapped in <span>

* Add integration test for loops with shapes and images inside them

* Add class method to ASTBuilder that checks for any block tags in a nodeset

This change also required moving the fetch_tag method to the class level.

* Only remove text nodes from root element if block level tags exist

This fixes a backwards compatabilty issue created by my refactor
of HTML insertion to be "inline" or "block".

* Remove span tag wrapping around "about_me" value in CV integration test