Support images (#90)
Adds support for images that are wrapped by a merge field block expression. Any image format can be used that MS Word supports as content types are added on the fly. The placeholder image is retained if the merge field expression evaluates to `nil`. 

* Create initial image content type

* Add method to DOM to add new media to /word/media

* Add Image operation to Sablon::Statement

* Add logic to support adding new content types to the docx file

* Add image block replacement logic to document processor

* Fix multiple image insertion

Insights pulled from https://github.com/pracstrat/sablon/commit/9112485e521680015eefe04d79181699a5a7c411
where they fixed the same issue in a slightly different way

* Update README with instructions for image support.

* Add image test fixtures

* Add document and context unit tests for images

* Fix bug in context transformation where hashes inside arrays did not get converted

This affected both keys that needed to be converted to strings and the
evaluation of special content type keys.

* Update Image handling logic to allow image reuse across XML files

i.e. an image in both the header and main document won't be duplicated.
Instead a new relationship is created that points towards that same file.

* Add full integration test case for image insertion

* Remove unused placeholder image fixture

* Support multiple input formats for images

The input format for images can be a simple path, a full URL to
a remote image or any object that implements a #read method such
as StringIO. If the object is "readable" then the filename must be
either passed in with the attributes, defined as a #filename method
on the "readable obj" or respond properly to File.basename.

* Add full content test for Images

* Add test case for when a filename can't be found and inspect string

This is important from a usability standpoint and helps
nail things down as "expected behavior".

* Improve README docs on other types of image data sources