Implement merge field handler registration (#99)
* Move all of the document block definitions into a separate file

This greatly improves the readability of the code since it helps
tidy up the main document class where the real work is done. The
namespaces are unchanged.

* Move operation construction logic into a separate file

This change was done in preparation for adding new field handler logic
otherwise the main document class would have become cluttered up again.
All of the namespacing is still unchanged.

* Refactor operation construction to allow "field handler" registration.

* Rename FieldHandler#process to FieldHandler#build_statement

This name is more representative of what the method is expected to
return.

* Tweak logic in document merge field handling to make it more testable

Essenitally, the only change was to use the #field_handlers and
in test cases. Additionally, it also prevents the edge case of the
@field_handlers variable being unitialized if someone would do odd
things in a monkey patch or fork.

* Write tests for document field handler methods

* Fix minor typos in sablon_test.rb

* Create integration test that adds new merge field handlers

This test adds a default handler for insertion so even fields without
an equals sign will get replaced and a new version of conditional blocks
where an equals/not equals operator is permitted.

* Use gsub on insertion expressions instead of slicing the string

This make it easier to support default insertion without an equals
sign since you don't need a custom InsertionHandler it just needs copied
over into the default spot. (see the custom_field_hander_test.rb file)