Implement a block reference counting scheme for merge fields
This reference counting scheme is used in multiblock contructs where
start and end fields get shared. The field cannot be removed until
only a single block depends on it. This would be better implemented on
the MergeField class itself that way the implementation of block doesn't
depend on it. But It gets tricky because sometimes only the field is
removed, other times it and it's parent node are removed.