Fix the custom field handler test case
The changes to my conditional operation broke the test. However,
I think this will be the last change to any existing conditionals.
due to the fact it breaks things when I cut a release it might be a
minor version increment instead of just a patch. A minor version
bump would also make sense because of the new field handler API
added.