Move all of the block reference checking logic to the MergeField classes
I still have to increment the block references during the intialization
method of Block but all of the ref reducing logic can be handled by
the MergeField instances. Since if I want to remove the field(s) then
clearly the block trying to remove them doesn't need to maintain a reference
to them.