Support full if elsif else blocks (#101)
* Update conditional statement to work with multiblock structures.

* Adds a method to consume multiple block based on sub expression patterns

This is only used for if/elseif/else blocks but in theory could be
expanded for any multi-block construct.

* Update the conditional field handler to build multiblock conditional statement

* Implement a block reference counting scheme for merge fields

This reference counting scheme is used in multiblock contructs where
start and end fields get shared. The field cannot be removed until
only a single block depends on it. This would be better implemented on
the MergeField class itself that way the implementation of block doesn't
depend on it. But It gets tricky because sometimes only the field is
removed, other times it and it's parent node are removed.

* Write unit tests for conditionals with elsif/else clauses in document_test [ci skip]

* Update sablon conditionals integration test to have if/elsif/else blocks

I test both block and inline forms. Behavior in tables wasn't directly
tested because of they use the same logic as paragraphs.

* Fix the custom field handler test case

The changes to my conditional operation broke the test. However,
I think this will be the last change to any existing conditionals.
due to the fact it breaks things when I cut a release it might be a
minor version increment instead of just a patch. A minor version
bump would also make sense because of the new field handler API
added.

* Begin generalizing block logic to clean things up

I need to make a few changes that will allow me to manage block
references in the merge fields easier.

* Remove special parent and body logic from ImageBlock

* Return only the first parent instead of the entire array in Block.parent

This type of logic will be used by the merge field class's "remove_parent"
method which is coming soon.

* Move all of the block reference checking logic to the MergeField classes

I still have to increment the block references during the intialization
method of Block but all of the ref reducing logic can be handled by
the MergeField instances. Since if I want to remove the field(s) then
clearly the block trying to remove them doesn't need to maintain a reference
to them.

* Update README with example for elsif/else blocks in conditionals