Implement basic dynamic image size support (#107)
Allows image sizes to be altered by providing the desired size in inches or centimeters. Aspect ratio is not locked when altering a single dimension, both `width` and `height` must be provided to maintain a given ratio. Pixels are not permitted because the true image size is dependent on the resolution of the provided image.

i.e. `Sablon.content(:image, filename, properties: { height: '1.0in', width: '1.5in' })`

* Implemented basic dynamic image size support
* CR fixes
* Simplify properties hash access for Image content object
* Allow floating point values as dimension strings
* Throw error when unit is not "cm" or "in"
* Update unit tests and integration tests