Implement logic to replace the parent of an InlineParagraphBlock
This block handles logic where the start and end fields occur inside
the same paragraph. This works without issue unless the content being
injected (i.e. WordML or HTML) is not allowed inside a Paragraph.
When this occured an undefined method `next_element' for nil:NilClass
exception would be throw because a parent paragraph could not be found.