Fix inline paragraph block level replace bug (#131)
* Implement logic to replace the parent of an InlineParagraphBlock

This block handles logic where the start and end fields occur inside
the same paragraph. This works without issue unless the content being
injected (i.e. WordML or HTML) is not allowed inside a Paragraph.
When this occured an undefined method `next_element' for nil:NilClass
exception would be throw because a parent paragraph could not be found.

* Add integration test for InlineParagraphBlock parent replacement

* Remove unneeded ParagraphBlock inheirentance from ImageBlock

Inheiriting from the Block class is sufficent, and not as misleading.

* Implement test to exercise inline image insertion

This includes inline if and loop constructs.